// src/front/project/src/types/PerformanceIndicators.ts
export interface PerformanceIndicators {
  vehiclesNeedingRepair: number;
  totalInspections: number;
  repairPercentage: number;
  averageAnalysisTime: number;
  vehicleUtilizationRate: number;
  totalVehicles: number;
  rentedVehicles: number;
  availableVehicles: number;
  maintenanceVehicles: number;
}